validação de CPF e garantia de inicidade  na base

+++ src/main/java/controller/ClientController.java
@@ -2,14 +2,19 @@ package controller;
 
 import java.text.ParseException;
 import java.util.List;
-import java.util.Optional;
+import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import Interface.ApoliceRepository;
 import Interface.ClienteRepository;
 import message.Response;
@@ -29,48 +34,51 @@ public class ClientController {
     
     @RequestMapping("/painel")
     public String WebCliente(Model model){
-    	model.addAttribute("clientes",  new Clientes());
+    	model.addAttribute("clientes", new Clientes());
     	selectAll();
        return "cliente";
     }
    
-    @RequestMapping("/salvar")
-    public String insert(@RequestParam String name, @RequestParam String city, @RequestParam String uf, @RequestParam String cpf) {
-        Clientes newCliente = new Clientes();
-        newCliente.setName(name);
-        newCliente.setCity(city);
-        newCliente.setUf(uf);
-        newCliente.setCpf(cpf);
-        clientesRepository.save(newCliente);
-        selectAll();
-        return "redirect:/clientes/painel";
+	@RequestMapping(value = "/novo", method = RequestMethod.GET)
+	public ModelAndView novo(Clientes cliente) {
+		ModelAndView mv = new ModelAndView("cliente");
+		return mv;
+	}
+	
+    @RequestMapping(value = {"/salvar"},  method = RequestMethod.POST)
+    public ModelAndView insert(@Valid @ModelAttribute("clientes") Clientes cliente,
+            BindingResult result, RedirectAttributes attributes ) {
+    	if(result.hasErrors())
+     	  return novo(cliente);
+    	else if ((clientesRepository.findCFP(cliente.getCpf()).size()) != 0) {
+    	  result.addError(new FieldError("clientes", "cpf", "CPF já está cadastrado!"));
+    	  return novo(cliente);
+    	}else
+    		clientesRepository.save(cliente);	
+    	ModelAndView mv = new ModelAndView("redirect:/clientes/novo");
+    	attributes.addFlashAttribute("mensagem", "Cliente salvo com sucesso.");
+        return mv;
     }
     
-    @RequestMapping("/editar")
-    public String update(@RequestParam String id, @RequestParam String name, @RequestParam String city, @RequestParam String uf,
-    		@RequestParam String cpf){
-        Optional<Clientes> cliente = clientesRepository.findById(id);
-        cliente.get().setName(name);
-        cliente.get().setCity(city);
-        cliente.get().setCpf(cpf);
-        cliente.get().setUf(uf);
-        clientesRepository.save(cliente.get());
-        return "redirect:/clientes/painel";
+    @RequestMapping(value= "/excluir",  method = RequestMethod.POST)
+    public ModelAndView delete(@Valid @ModelAttribute("clientes") Clientes cliente,
+            BindingResult result, RedirectAttributes attributes) {
+    	ModelAndView mv = new ModelAndView("redirect:/clientes/novo");
+    	Clientes _rCliente = clientesRepository.findById(cliente.getId()).get();
+        List<Apolice> contratosApolices = apoliceRepository.findContrato(_rCliente.getId());
+        if(cliente.getOperacao().equals("EXCLUIR")) {
+	    	clientesRepository.delete(_rCliente); 
+	    	for (Apolice apolices : contratosApolices) { // Exclui todos contratos reencindidos associados ao cliente que foi excluido
+	    	  apoliceRepository.delete(apolices);
+			}
+	    	attributes.addFlashAttribute("mensagem", "Cliente excluído com sucesso.");
+        }else {
+        	clientesRepository.save(_rCliente);	
+        	attributes.addFlashAttribute("mensagem", "Cadastro de cliente editado com sucesso.");
+        }
+    	return mv;
     }
-
-    @RequestMapping("/excluir/{id}")
-    @ResponseBody
-    public Response delete(@PathVariable String id) {
-    	  Clientes cliente = clientesRepository.findById(id).get();
-        List<Apolice> contratosApolices = apoliceRepository.findContrato(cliente.getId());
-        Response response = new Response("Done", cliente);
-    	  clientesRepository.delete(cliente); 
-    	  for (Apolice apolices : contratosApolices) { // Exclui todos contratos reencindidos associados ao cliente que foi excluido
-    	     apoliceRepository.delete(apolices);
-				}
-    	return response;
-    }
-	
+    
     @RequestMapping("/buscarTodos")
     @ResponseBody
     public Response selectAll() {
